Pull reservation question/answer tests into its own system test file
